chore(api): move obligation endpoints from license
